backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
django-cors-headers==4.3.1
python-multipart==0.0.6
Pillow==10.1.0
openai-whisper==20231117
torch==2.1.1
torchaudio==2.1.1
moviepy==1.0.3
language-tool-python==2.7.1
nltk==3.8.1
opencv-python==4.8.1.78
numpy==1.26.4
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
